# cli/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"XReserveTokenInfo": {
				"properties": {
					"bridgeAddress": {
						"type": "string"
					},
					"feeConst": {
						"type": "string"
					},
					"feeShare": {
						"type": "string"
					},
					"protocolAddress": {
						"type": "string"
					}
				},
				"required": [
					"bridgeAddress",
					"feeConst",
					"feeShare"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuiAddresses": {
				"properties": {
					"bridgeAddress": {
						"type": "string"
					},
					"bridgeAddressOrigin": {
						"type": "string"
					},
					"bridgeObjectAddress": {
						"type": "string"
					},
					"allbridgeMessengerAddress": {
						"type": "string"
					},
					"allbridgeMessengerAddressOrigin": {
						"type": "string"
					},
					"allbridgeMessengerObjectAddress": {
						"type": "string"
					},
					"gasOracleAddress": {
						"type": "string"
					},
					"gasOracleAddressOrigin": {
						"type": "string"
					},
					"gasOracleObjectAddress": {
						"type": "string"
					},
					"utilsAddress": {
						"type": "string"
					},
					"wormholeMessengerAddress": {
						"type": "string"
					},
					"wormholeMessengerAddressOrigin": {
						"type": "string"
					},
					"wormholeMessengerObjectAddress": {
						"type": "string"
					},
					"wormholeStateObjectAddress": {
						"type": "string"
					},
					"cctpAddress": {
						"type": "string"
					},
					"cctpAddressOrigin": {
						"type": "string"
					},
					"cctpObjectAddress": {
						"type": "string"
					},
					"cctpDenyListObjectAddress": {
						"type": "string"
					},
					"cctpMessageTransmitterStateObjectAddress": {
						"type": "string"
					},
					"cctpTokenMessengerMinterStateObjectAddress": {
						"type": "string"
					},
					"cctpTreasuryObjectAddress": {
						"type": "string"
					}
				},
				"required": [
					"bridgeAddress",
					"bridgeObjectAddress",
					"allbridgeMessengerAddress",
					"allbridgeMessengerObjectAddress",
					"gasOracleAddress",
					"gasOracleObjectAddress",
					"utilsAddress",
					"wormholeMessengerAddress",
					"wormholeMessengerObjectAddress",
					"wormholeStateObjectAddress",
					"cctpAddress",
					"cctpObjectAddress",
					"cctpDenyListObjectAddress",
					"cctpMessageTransmitterStateObjectAddress",
					"cctpTokenMessengerMinterStateObjectAddress",
					"cctpTreasuryObjectAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChainSymbol": {
				"enum": [
					"BSC",
					"ETH",
					"BAS",
					"SOL",
					"TRX",
					"POL",
					"ARB",
					"CEL",
					"AVA",
					"SRB",
					"STLR",
					"OPT",
					"SUI",
					"SNC",
					"UNI",
					"LIN",
					"ALG",
					"STX"
				],
				"type": "string"
			},
...
			"AbrPayerChainInfo": {
				"properties": {
					"payerAddress": {
						"type": "string",
						"description": "ABR proxy for paying fees with ABR"
					},
					"abrToken": {
						"$ref": "#/components/schemas/TokenCoreFields"
					},
					"payerAvailability": {
						"$ref": "#/components/schemas/AbrPayerAvailability",
						"description": "Availability to pay by abr"
					}
				},
				"required": [
					"payerAddress",
					"abrToken",
					"payerAvailability"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TransferTime": {
				"$ref": "#/components/schemas/Record_string.MessengerTransferTime_"
			},
			"TxCostAmount": {
				"description": "Contains transfer costs",
				"properties": {
					"maxAmount": {
						"type": "string",
						"description": "The maximum gas amount that can be received as extra gas for one transfer"
					},
					"swap": {
						"type": "string",
						"description": "Swap cost"
					},
					"transfer": {
						"type": "string",
						"description": "Transfer cost"
					}
				},
				"required": [
					"maxAmount",
					"swap",
					"transfer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChainType": {
				"enum": [
					"EVM",
					"SOLANA",
					"TRX",
					"SRB",
					"SUI",
					"ALG",
					"STX"
				],
				"type": "string"
			},
...
			"TokenWithChainDetails": {
				"description": "General Token Interface",
				"properties": {
					"symbol": {
						"type": "string",
						"description": "Token symbol"
					},
					"name": {
						"type": "string",
						"description": "Token name"
					},
					"decimals": {
						"type": "number",
						"format": "double",
						"description": "Token decimals"
					},
					"poolAddress": {
						"type": "string",
						"description": "Token pool address"
					},
					"tokenAddress": {
						"type": "string",
						"description": "Token address"
					},
					"originTokenAddress": {
						"type": "string",
						"description": "Token origin address"
					},
					"cctpAddress": {
						"type": "string",
						"description": "Token CCTP address</br>\nOptional. Defined if CCTP supported by token"
					},
					"cctpFeeShare": {
						"type": "string",
						"description": "Token CCTP V2 address</br>\nOptional. Defined if CCTP supported by token"
					},
					"cctpV2Address": {
						"type": "string",
						"description": "Token CCTP address</br>\nOptional. Defined if CCTP supported by token"
					},
					"cctpV2FeeShare": {
						"type": "string",
						"description": "Token CCTP V2 address</br>\nOptional. Defined if CCTP supported by token"
					},
					"xReserve": {
						"$ref": "#/components/schemas/XReserveTokenInfo",
						"description": "XReserve bridge configuration for token.\nOptional. Defined if XReserve supported by token"
					},
					"oftId": {
						"type": "string",
						"description": "Internal identifier for the same token across different chains.\n\nUsed to link representations of a token deployed on multiple networks (e.g., Ethereum, BSC).\nTokens with the same `oftId` are considered to be equivalent.\n\nOptional. Defined if OFT supported by token"
					},
					"feeShare": {
						"type": "string",
						"description": "Token fee share"
					},
					"apr": {
						"type": "string"
					},
					"apr7d": {
						"type": "string",
						"description": "Token APR based on last 7 days"
					},
					"apr30d": {
						"type": "string",
						"description": "Token APR based on last 30 days"
					},
					"lpRate": {
						"type": "string",
						"description": "Token LP rate"
					},
					"yieldId": {
						"type": "number",
						"format": "double"
					},
					"suiAddresses": {
						"$ref": "#/components/schemas/SuiAddresses",
						"description": "Sui addresses\nOptional. Defined for SUI"
					},
					"allbridgeChainId": {
						"type": "number",
						"format": "double",
						"description": "Allbridge's Id"
					},
					"bridgeId": {
						"type": "string",
						"description": "Algo Bridge appId"
					},
					"paddingUtilId": {
						"type": "string",
						"description": "Algo Padding appId"
					},
					"bridgeAddress": {
						"type": "string",
						"description": "Bridge contract address"
					},
					"oftBridgeAddress": {
						"type": "string",
						"description": "OFT Bridge contract address\nOptional. Defined if the chain supports OFT"
					},
					"yieldAddress": {
						"type": "string",
						"description": "Yield contract address"
					},
					"abrPayer": {
						"$ref": "#/components/schemas/AbrPayerChainInfo",
						"description": "ABR payer chain info"
					},
					"transferTime": {
						"$ref": "#/components/schemas/TransferTime",
						"description": "Average transfer time to other blockchains"
					},
					"txCostAmount": {
						"$ref": "#/components/schemas/TxCostAmount",
						"description": "Transfers costs"
					},
					"confirmations": {
						"type": "number",
						"format": "double",
						"description": "Number of confirmations required"
					},
					"chainSymbol": {
						"$ref": "#/components/schemas/ChainSymbol",
						"description": "The symbol of the chain representing one of the supported blockchain networks (e.g., \"ETH\" for Ethereum). For more details, see: {@link ChainSymbol}."
					},
					"chainId": {
						"type": "string",
						"description": "Common Blockchain Id</br>\nA 0x-prefixed hexadecimal string</br>\nOptional."
					},
					"chainType": {
						"$ref": "#/components/schemas/ChainType",
						"description": "Blockchain type. For more details, see: {@link ChainType}."
					},
					"chainName": {
						"type": "string",
						"description": "Blockchain network name"
					}
				},
				"required": [
					"symbol",
					"name",
					"decimals",
					"poolAddress",
					"tokenAddress",
					"feeShare",
					"apr",
					"apr7d",
					"apr30d",
					"lpRate",
					"allbridgeChainId",
					"bridgeAddress",
					"transferTime",
					"txCostAmount",
					"confirmations",
					"chainSymbol",
					"chainType",
					"chainName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"HttpExceptionBody": {
				"properties": {
					"message": {
						"$ref": "#/components/schemas/HttpExceptionBodyMessage"
					},
					"error": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"YieldWithdrawAmount": {
				"properties": {
					"amount": {
						"type": "string",
						"description": "The float amount of tokens to be withdrawn."
					},
					"token": {
						"$ref": "#/components/schemas/TokenWithChainDetailsYield"
					}
				},
				"required": [
					"amount",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RawTransaction": {
				"$ref": "#/components/schemas/SdkRawTransaction"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "3.30.0-alpha.2",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/yield/tokens": {
			"get": {
				"operationId": "GetYieldTokens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TokenWithChainDetails"
									},
									"type": "array"
								},
								"examples": {
									"Available yield tokens": {
										"value": [
											{
												"chainSymbol": "ETH",
												"chainId": "1",
												"name": "USD Coin",
												"symbol": "USDC",
												"decimals": 6,
												"tokenAddress": "0x0000000000000000000000000000000000000001",
												"yieldAddress": "0x0000000000000000000000000000000000000002",
												"tokens": [
													{
														"chainSymbol": "ETH",
														"chainId": "1",
														"name": "USD Coin",
														"symbol": "USDC",
														"decimals": 6,
														"tokenAddress": "0x0000000000000000000000000000000000000001",
														"yieldAddress": "0x0000000000000000000000000000000000000002",
														"yieldId": "1"
													}
												]
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens",
					"Yield"
				],
				"security": [],
				"parameters": []
			}
		},
		"/yield/allowance": {
			"get": {
				"operationId": "GetYieldAllowance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Yield allowance amount": {
										"value": "100000000"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Yield",
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/check/yield/allowance": {
			"get": {
				"operationId": "CheckYieldAllowance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								},
								"examples": {
									"Yield allowance check result": {
										"value": true
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens",
					"Yield"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/yield/balance": {
			"get": {
				"operationId": "GetYieldTokenBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "string"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								},
								"examples": {
									"Yield token balance in smallest units": {
										"value": {
											"result": "2500000"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens",
					"Yield"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/yield/deposit/calculate": {
			"get": {
				"operationId": "GetYieldAmountToBeDeposited",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Estimated CYD amount on deposit": {
										"value": "2.4987"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Yield"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/yield/withdrawn/calculate": {
			"get": {
				"operationId": "GetYieldAmountToBeWithdrawn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/YieldWithdrawAmount"
									},
									"type": "array"
								},
								"examples": {
									"Estimated withdrawn token amounts": {
										"value": [
											{
												"amount": "1.25",
												"token": {
													"chainSymbol": "ETH",
													"chainId": "1",
													"name": "USD Coin",
													"symbol": "USDC",
													"decimals": 6,
													"tokenAddress": "0x0000000000000000000000000000000000000001",
													"yieldAddress": "0x0000000000000000000000000000000000000002",
													"yieldId": "1"
												}
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Yield"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "yieldAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/raw/yield/approve": {
			"get": {
				"operationId": "ApproveYield",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Yield approve raw transaction": {
										"value": "0xffff"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Yield",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "amount",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/raw/yield/deposit": {
			"get": {
				"operationId": "DepositYield",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Yield deposit raw transaction": {
										"value": "0xffff"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Yield",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minVirtualAmount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/raw/yield/withdraw": {
			"get": {
				"operationId": "WithdrawYield",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Yield withdraw raw transaction": {
										"value": "0xffff"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Yield",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "yieldAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/raw/approve": {
			"get": {
				"operationId": "Approve",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"EVM approve transaction": {
										"value": {
											"from": "0x0000000000000000000000000000000000000000",
											"to": "0x0000000000000000000000000000000000000000",
											"value": "1000",
											"data": "0x0000000000000000000000000000000000000000000000000000000000000000"
										}
									},
									"Tron approve transaction": {
										"value": {
											"visible": false,
											"txID": "0000000000000000000000000000000000000000000000000000000000000000",
											"raw_data": {
												"contract": [],
												"ref_block_bytes": "0000",
												"ref_block_hash": "0000000000000000",
												"expiration": 0,
												"timestamp": 0
											},
											"raw_data_hex": "00"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Transfers",
					"Raw Transactions",
					"Yield"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "amount",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"default": "bridge",
							"type": "string",
							"enum": [
								"bridge",
								"pool",
								"yield"
							]
						}
					},
					{
						"in": "query",
						"name": "messenger",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ALLBRIDGE",
								"WORMHOLE",
								"CCTP",
								"CCTP_V2",
								"OFT",
								"X_RESERVE"
							]
						}
					},
					{
						"in": "query",
						"name": "feePaymentMethod",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"WITH_NATIVE_CURRENCY",
								"WITH_STABLECOIN",
								"WITH_ABR"
							]
						}
					},
					{
						"in": "query",
						"name": "contractAddress",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/raw/deposit": {
			"get": {
				"operationId": "Deposit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Pool deposit raw transaction": {
										"value": "0xffff"
									},
									"Stacks pool deposit transaction": {
										"value": "80800000000400b168de3a000000000000000100000000000000000000000000000000"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeParams",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"AUTO",
								"PRICE_PER_UNIT_IN_MICRO_LAMPORTS",
								"EXTRA_FEE_IN_LAMPORTS"
							]
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeValue",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/raw/withdraw": {
			"get": {
				"operationId": "Withdraw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Pool withdraw raw transaction": {
										"value": "0xffff"
									},
									"Stacks pool withdraw transaction": {
										"value": "80800000000400b168de3a000000000000000100000000000000000000000000000000"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeParams",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"AUTO",
								"PRICE_PER_UNIT_IN_MICRO_LAMPORTS",
								"EXTRA_FEE_IN_LAMPORTS"
							]
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeValue",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}